+++ app/src/main/java/com/techcourse/DispatcherServletInitializer.java
@@ -4,7 +4,11 @@ import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import web.org.springframework.web.WebApplicationInitializer;
-import webmvc.org.springframework.web.servlet.DispatcherServlet;
+import webmvc.org.springframework.web.servlet.mvc.DispatcherServlet;
+import webmvc.org.springframework.web.servlet.mvc.tobe.adapter.HandlerAdapters;
+import webmvc.org.springframework.web.servlet.mvc.tobe.handler.HandlerMappings;
+import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotaionHandlerAdapter;
+import webmvc.org.springframework.web.servlet.mvc.tobe.AnnotationHandlerMapping;
 
 /**
  * Base class for {@link WebApplicationInitializer}
@@ -18,7 +22,13 @@ public class DispatcherServletInitializer implements WebApplicationInitializer {
 
     @Override
     public void onStartup(final ServletContext servletContext) {
-        final var dispatcherServlet = new DispatcherServlet();
+
+        final HandlerMappings handlerMappings = new HandlerMappings();
+        handlerMappings.add(new AnnotationHandlerMapping("com.mvc"));
+        final HandlerAdapters handlerAdapters = new HandlerAdapters();
+        handlerAdapters.add(new AnnotaionHandlerAdapter());
+
+        final var dispatcherServlet = new DispatcherServlet(handlerMappings, handlerAdapters);
 
         final var registration = servletContext.addServlet(DEFAULT_SERVLET_NAME, dispatcherServlet);
         if (registration == null) {
